Implemented drawing of map
Added Level definition, with specification for map of first level.
Map is not yet moving.
Revisited initial implementation to avoid use of dynamically allocated memory. This does not work well very limited memory that is available on Arduino chip.

Disabled iso-line rendering for now. This will have to be refactored.
Also disabled use of MapUnit. This also needs refactoring.