Improved performance of tile drawing
There's a significant difference in speed required for drawing an image with transparency and images without. Drawing images with transparency is nearly twice is slow. More specifically, in my benchmarks it's 1.75 times as slow.

Therefore changed rendering to only use transparency at the top of tile, where it is needed to prevent drawing over the tiles behind it. The remainder of the tile is drawn without transparency. An additional benefit is that this can also be used to give the tiles at the front of the map clean borders.

The impact of these changes is significant. At 15 FPS, CPU usage for Level 1 went down from 86% to 63%. This difference is even bigger than it looks like, as an empty loop (at 25 FPS) already uses 50% of the CPU, "just" for refreshing the display.

With these improvements, 20 FPS definitely looks feasible for this game, and possibly even 25 FPS. Also note that CPU usage for Level 0 is lower than for Level 1, even though the former uses bigger images. Apparently the overhead of calculating the waves is significant.