Support movers moving off the map
Extended TilePos to represent the necessary off-map positions.
Added a single off-map tile to handle drawing of the off-map mover. This should always be one at most at any time (as only pushed boxes and players will be able to move off the map).

Note: The move-behavior is undefined when a mover moves when it is off-the-map. The mover will re-appear at an arbitrary position. This is not a problem, as this can not happen anymore once the player dies when falling off the map.