Minor tweaks to timing
Slowed down bots to make game more playable.
Adapted waves to that with the waves, in both column and row direction, it is possible to cross the entire map without bumping. Against the waves, bots can only move two tiles at once. This resembles the dynamics of the original PICO-8 game.