Minor fix to rounding for Mover DX/DY values
Without this fix, not all Y values were chosen equally often. All values should be chosen twice (when movementDelay is one). However, when the mover switched drawing tile, the Y-value was the same three times in a row. This resulted in a noticeable "bounce" while moving. It also meant the mover sprite was partially obscured by the front tile.