Improved camera movement
When level is shown, zoom in onto center of tiles.
When level starts (i.e. when movers are added), zoom in to player.
When level is frozen, keep camera on player.

Also changed positioning so that player is really in the center of the screen (unless he's near the edge)