Fixed bug that could cause mover to fall through the map
When reversing at the right (wrong) moment, just after exitedTile has been called, the Mover would drop of the map because enteringTile was not called, so swapUnits would remove it from the map.
This fixes that, and is how the issue was handled in the original PICO-8 code. It's a bit ugly though.