Prevent mover from moving while falling
Changes of commit 312e6ea did not suffice.
Now setting falling status when mover exited its start tile.
This way, the falling status can be used to prevent starting a new move, as well as not allowing retreats (while move is still ongoing).
It still allows the player to retreat when it is about to fall but still on two tiles.