Fixed drawing of dropping movers
Now only drawing the top-part of the mover that is not yet inside the gap.
To do so added drawDroppingImage function, which is based on drawing function in Gamebuino library, but extended to support masking.

Also moved dropping logic to Mover base class. The reason that enemies will not drop into gaps is not enough reason to not support it there. This way, no need to replicate dropping logic in both Player and Box subclasses.