Improved stats screen
Use three colors for progress bar: completed, pending (but unlocked) and locked.
With this, reduced size back to sixteen cells, one for each level.
These changes ensure that after game completion the progress bar is entirely green, which it was not before.

Also changed look of GAME and BEST headers. Changed text to black on colored background.